+++ delivery-server/src/main/java/chill_logistics/delivery_server/application/service/AsyncAiService.java
@@ -1,4 +1,4 @@
-package chill_logistics.delivery_server.application;
+package chill_logistics.delivery_server.application.service;
 
 import chill_logistics.delivery_server.application.dto.command.HubRouteAfterCommandV1;
 import chill_logistics.delivery_server.application.dto.command.HubRouteHubInfoV1;
