+++ delivery-server/src/main/java/chill_logistics/delivery_server/application/service/AsyncDiscordService.java
@@ -1,4 +1,4 @@
-package chill_logistics.delivery_server.application;
+package chill_logistics.delivery_server.application.service;
 
 import chill_logistics.delivery_server.infrastructure.ai.dto.request.AiDeadlineRequestV1;
 import chill_logistics.delivery_server.infrastructure.ai.dto.response.AiDeadlineResponseV1;
